Implement ChromaDbVectorStore for document storage and retrieval; update vector store interface and refactor related components
